Changed precision of share values from int to double.

+++ src/common/rmi/RemoteTransactions.java
@@ -17,9 +17,9 @@ import java.util.ArrayList;
  */
 public interface RemoteTransactions extends Remote
 {
-	public void setShareValue(int user_id, int share_id, int new_value) throws RemoteException, SQLException;
+	public void setShareValue(int user_id, int share_id, double new_value) throws RemoteException, SQLException;
 
-	public int buyShares(int user_id, int idea_id, int share_num, int price_per_share, int new_price_share, boolean fromQueue) throws RemoteException, SQLException, NotEnoughCashException, NotEnoughSharesException;
+	public int buyShares(int user_id, int idea_id, int share_num, double price_per_share, double new_price_share, boolean fromQueue) throws RemoteException, SQLException, NotEnoughCashException, NotEnoughSharesException;
 
 	public ArrayList<ShareInfo> getShares(int idea_id) throws RemoteException, SQLException;
